+++ packages/web3-providers/tests/src/providers/WebsocketProviderTest.js
@@ -22,8 +22,8 @@ describe('WebsocketProviderTest', () => {
         expect(websocketProvider.connection).toEqual(socketMock);
 
         expect(websocketProvider.timeout).toEqual(1);
-        
-        expect(websocketProvider.reconnectionTimeout).toEqual(1);
+
+        expect(websocketProvider.reconnectDelay).toEqual(1);
 
         expect(socketMock.addEventListener.mock.calls[0][0]).toEqual('message');
         expect(socketMock.addEventListener.mock.calls[0][1]).toBeInstanceOf(Function);
@@ -450,7 +450,6 @@ describe('WebsocketProviderTest', () => {
 
     it('calls onConnect after the connection got lost', (done) => {
         websocketProvider.on('reconnected', () => {
-
             done();
         });