+++ CHANGELOG.md
@@ -66,12 +66,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Length check of the PK added to the ``fromPrivateKey`` method of the ``Account`` model (#2928)
-- WebsocketProvider options extended with ``requestOptions`` (#2938)
+- WebsocketProvider options extended with ``requestOptions`` and ``origins`` (#2938, #2995)
 - ``changed`` listener added to Contract event subscriptions (#2960)
 
 ### Changed
 
 - fsevents bumbed to v1.2.9 (#2951)
+- ``websocket`` dependency changed to github fork (#2995)
 
 ### Fixed